Removed obsolete standalone support for ASL/os_log
