--- Source/Generic/ECOSLogClient.h
@@ -1,12 +1,0 @@
-// --------------------------------------------------------------------------
-//
-//  Copyright 2014 Sam Deane, Elegant Chaos. All rights reserved.
-//  This source code is distributed under the terms of Elegant Chaos's
-//  liberal license: http://www.elegantchaos.com/license/liberal
-// --------------------------------------------------------------------------
-
-#import "ECSystemLogClient.h"
-
-@interface ECOSLogClient : ECSystemLogClient
-
-@end
